src/requirements.txt
streamlit==1.26.0
boto3==1.34.54
botocore==1.34.45
yfinance==0.2.34
pandas_ta==0.3.16b0
numpy==1.25.1
requests==2.31.0
aiohttp==3.9.7
aioboto3==11.0.0
asyncio==3.4.3
python-dotenv==1.1.0
langchain-aws==0.1.2
langchain-community==1.1.6
singleton_decorator==1.0.0